renamed logging_api to peripheral_api
