+++ api/config/config.example.json
@@ -26,7 +26,7 @@
     "SCE_EVENTS_CAL": "XXXXXXXXXXXXXXXXXXXXXXXXXXX"
   },
   "ledSignUrl": "XXXXXXX",
-  "LOGGING_API_URL": "http://localhost:8081/loggingapi",
+  "PERIPHERAL_API_URL": "http://localhost:8081/loggingapi",
   "passwordStrength": "medium",
   "S3Bucket": {
     "AWSACCESSKEYID": "XXXXXXXXXXXXXXXXXXXX",
@@ -46,7 +46,7 @@
   "PrintingS3Bucket": "XXXXXXXXXXXXXXX",
   "secretKey": "super duper secret key",
   "GENERAL_API_URL_PROD": "http://mainendpoints:8080/api",
-  "LOGGING_API_URL_PROD": "http://sce-logging-api:8081/loggingapi",
+  "PERIPHERAL_API_URL_PROD": "http://sce-logging-api:8081/loggingapi",
   "MAILER_API_URL_PROD": "http://sce-cloud-api:8082/cloudapi",
   "RPC_API_URL_PROD": "http://localhost:8083/SceRpcApi"
 }
