+++ api/devServer.js
@@ -5,6 +5,6 @@ if (typeof module !== 'undefined' && !module.parent) {
   // Starting servers
   require('./main_endpoints/server');
   require('./cloud_api/server');
-  require('./logging_api/server');
+  require('./peripheral_api/server');
 }
