+++ api/peripheral_api/routes/PrintLog.js
@@ -1,5 +1,5 @@
 
-   
+
 const express = require('express');
 const router = express.Router();
 const axios = require('axios');
