+++ src/APIFunctions/2DPrinting.js
@@ -8,15 +8,15 @@ import {
 let config = require('../config/config.json');
 let GENERAL_API_URL = process.env.NODE_ENV === 'production' ?
   config.GENERAL_API_URL_PROD : config.GENERAL_API_URL;
-let LOGGING_API_URL = process.env.NODE_ENV === 'production' ?
-  config.LOGGING_API_URL_PROD : config.LOGGING_API_URL;
+let PERIPHERAL_API_URL = process.env.NODE_ENV === 'production' ?
+  config.PERIPHERAL_API_URL_PROD : config.PERIPHERAL_API_URL;
 let RPC_API_URL = process.env.NODE_ENV === 'production' ?
   config.RPC_API_URL_PROD : config.RPC_API_URL;
 
   console.log({
 
     config,
-    LOGGING_API_URL
+    PERIPHERAL_API_URL
   });
 /**
  * Return an array similar to python's range() function
@@ -33,7 +33,7 @@ export const range = (start, end) => {
  */
 export async function healthCheck() {
   let status = new ApiResponse();
-  await axios.get(LOGGING_API_URL + '/Printer/healthCheck')
+  await axios.get(PERIPHERAL_API_URL + '/Printer/healthCheck')
     .then(res => {
       status.reponseData = res.data;
     })
@@ -88,7 +88,7 @@ export function parseRange(pages, maxPages) {
 export async function printPage(data) {
   console.log({data})
   let status = new ApiResponse();
-  await axios.post(LOGGING_API_URL + '/Printer/sendPrintRequest', data)
+  await axios.post(PERIPHERAL_API_URL + '/Printer/sendPrintRequest', data)
     .then(response => {
       status.responseData = response.data.message;
     })
@@ -106,7 +106,7 @@ export async function printPage(data) {
  */
 export async function logPrintRequest(data) {
   let status = new ApiResponse();
-  await axios.post(LOGGING_API_URL + '/PrintLog/addPrintLog', data)
+  await axios.post(PERIPHERAL_API_URL + '/PrintLog/addPrintLog', data)
     .catch(() => {
       status.error = true;
     });
@@ -119,7 +119,7 @@ export async function logPrintRequest(data) {
  */
 export async function getAllLogs() {
   let status = new ApiResponse();
-  await axios.get(LOGGING_API_URL + '/PrintLog/getPrintLogs')
+  await axios.get(PERIPHERAL_API_URL + '/PrintLog/getPrintLogs')
     .then(response => {
       status.responseData = response.data;
     })
