+++ src/APIFunctions/LedSign.js
@@ -1,8 +1,8 @@
 import axios from 'axios';
 import { ApiResponse } from './ApiResponses';
 let config = require('../config/config.json');
-let LOGGING_API_URL = process.env.NODE_ENV === 'production' ?
-  config.LOGGING_API_URL_PROD : config.LOGGING_API_URL;
+let PERIPHERAL_API_URL = process.env.NODE_ENV === 'production' ?
+  config.PERIPHERAL_API_URL_PROD : config.PERIPHERAL_API_URL;
 
 /**
  * Checks to see if the sign is accepting requests. This is done
@@ -13,7 +13,7 @@ let LOGGING_API_URL = process.env.NODE_ENV === 'production' ?
 export async function healthCheck(officerName) {
   let status = new ApiResponse();
   await axios
-    .get(LOGGING_API_URL + '/LedSign/healthCheck', { officerName })
+    .get(PERIPHERAL_API_URL + '/LedSign/healthCheck', { officerName })
     .then(res => {
       status.responseData = res.data;
     })
@@ -32,7 +32,7 @@ export async function healthCheck(officerName) {
 export async function getAllSignLogs() {
   let result = new ApiResponse();
   await axios
-    .get(LOGGING_API_URL+'/SignLog/getSignLogs')
+    .get(PERIPHERAL_API_URL+'/SignLog/getSignLogs')
     .then(res => {
       result.responseData = res.data;
     })
@@ -53,7 +53,7 @@ export async function getAllSignLogs() {
 export async function updateSignText(signData) {
   let status = new ApiResponse();
   await axios
-    .post(LOGGING_API_URL + '/LedSign/updateSignText', { ...signData })
+    .post(PERIPHERAL_API_URL + '/LedSign/updateSignText', { ...signData })
     .then(res => {
       status = res.data;
     })
