+++ src/config/config.example.json
@@ -1,11 +1,11 @@
 {
   "GENERAL_API_URL": "http://localhost:8080/api",
-  "LOGGING_API_URL": "http://localhost:8081/loggingapi",
+  "PERIPHERAL_API_URL": "http://localhost:8081/loggingapi",
   "MAILER_API_URL": "http://localhost:8082/cloudapi",
   "RPC_API_URL": "http://localhost:8083/SceRpcApi",
   "DISCORD_SJSU_API_URL": "http://localhost:8080/api",
   "GENERAL_API_URL_PROD": "https://sce.engr.sjsu.edu/api",
-  "LOGGING_API_URL_PROD": "https://sce.engr.sjsu.edu/loggingapi",
+  "PERIPHERAL_API_URL_PROD": "https://sce.engr.sjsu.edu/loggingapi",
   "MAILER_API_URL_PROD": "https://sce.engr.sjsu.edu/cloudapi",
   "RPC_API_URL_PROD": "https://sce.engr.sjsu.edu/SceRpcApi",
   "DISCORD_SJSU_API_URL_PROD": "http://XXXXXXXXXXXX.XXX:8080/api",
