+++ test/PrintLog.js
@@ -1,7 +1,7 @@
 /* global describe it before after */
 // During the test the env variable is set to test
 process.env.NODE_ENV = 'test';
-const PrintLog = require('../api/logging_api/models/PrintLog');
+const PrintLog = require('../api/peripheral_api/models/PrintLog');
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 const constants = require('../api/util/constants');
@@ -18,7 +18,7 @@ chai.use(chaiHttp);
 describe('PrintLog', () => {
   before(done => {
     app = tools.initializeServer(
-      __dirname + '/../api/logging_api/routes/PrintLog.js');
+      __dirname + '/../api/peripheral_api/routes/PrintLog.js');
     test = new SceApiTester(app);
     tools.emptySchema(PrintLog);
     done();
